requirements.txt
groq
langchain
langchain-community
langchain-core
langchain-groq
faiss-cpu
tiktoken
pypdf
streamlit
python-dotenv
sentence-transformers
scikit-learn
